Rename Console.h/cpp to Logging and fix includes
