+++ src/Diagnostics/include/OpenLoco/Diagnostics/Logging.h
