+++ src/OpenLoco/src/Audio/Channel.cpp
@@ -1,5 +1,5 @@
 #include "Channel.h"
-#include <OpenLoco/Console/Console.h>
+#include <OpenLoco/Diagnostics/Logging.h>
 #include <utility>
 
 namespace OpenLoco::Audio
