+++ src/OpenLoco/src/Graphics/Gfx.cpp
@@ -12,7 +12,7 @@
 #include "PaletteMap.h"
 #include "Ui.h"
 #include "Ui/WindowManager.h"
-#include <OpenLoco/Console/Console.h>
+#include <OpenLoco/Diagnostics/Logging.h>
 #include <OpenLoco/Interop/Interop.hpp>
 #include <OpenLoco/Utility/Stream.hpp>
 #include <algorithm>
