+++ src/OpenLoco/src/Localisation/StringManager.cpp
@@ -8,7 +8,7 @@
 #include "Objects/ObjectManager.h"
 #include "StringIds.h"
 #include "World/TownManager.h"
-#include <OpenLoco/Console/Console.h>
+#include <OpenLoco/Diagnostics/Logging.h>
 #include <OpenLoco/Interop/Interop.hpp>
 
 #include <cassert>
