+++ src/OpenLoco/src/Ui.cpp
@@ -44,7 +44,7 @@
 #include "Ui/WindowManager.h"
 #include "Window.h"
 #include "World/CompanyManager.h"
-#include <OpenLoco/Console/Console.h>
+#include <OpenLoco/Diagnostics/Logging.h>
 #include <OpenLoco/Interop/Interop.hpp>
 #include <OpenLoco/Utility/String.hpp>
