+++ src/OpenLoco/src/Window.cpp
@@ -11,7 +11,7 @@
 #include "Ui.h"
 #include "Ui/ScrollView.h"
 #include "Widget.h"
-#include <OpenLoco/Console/Console.h>
+#include <OpenLoco/Diagnostics/Logging.h>
 #include <OpenLoco/Engine/Ui/Rect.hpp>
 #include <OpenLoco/Interop/Interop.hpp>
 #include <OpenLoco/Utility/Numeric.hpp>
